requirements.txt
newspaper4k
lxml_html_clean
typing_extensions